Add dual chain support for state router, prevent update when invoke (#10691)
* Add dual chain support for state router, prevent update when invoke

* Fix uts
