+++ dubbo-cluster/src/main/java/org/apache/dubbo/rpc/cluster/router/state/AbstractStateRouter.java
@@ -113,15 +113,16 @@ public abstract class AbstractStateRouter<T> implements StateRouter<T> {
         }
         BitList<Invoker<T>> routeResult;
 
+        routeResult = doRoute(invokers, url, invocation, needToPrintMessage, nodeHolder, messageHolder);
+        if (routeResult != invokers) {
+            routeResult = invokers.and(routeResult);
+        }
         // check if router support call continue route by itself
         if (!supportContinueRoute()) {
-            routeResult = doRoute(invokers, url, invocation, needToPrintMessage, nodeHolder, messageHolder);
             // use current node's result as next node's parameter
             if (!shouldFailFast || !routeResult.isEmpty()) {
                 routeResult = continueRoute(routeResult, url, invocation, needToPrintMessage, nodeHolder);
             }
-        } else {
-            routeResult = doRoute(invokers, url, invocation, needToPrintMessage, nodeHolder, messageHolder);
         }
 
         // post-build current node
@@ -190,7 +191,7 @@ public abstract class AbstractStateRouter<T> implements StateRouter<T> {
     @Override
     public final String buildSnapshot() {
         return doBuildSnapshot() +
-            "            ↓ \n" +
+            "            v \n" +
             nextRouter.buildSnapshot();
     }
